moved test stream pool creation to preload library to override create_global_cuda_stream_pool
